+++ include/unigui/widgets/window.h
@@ -20,6 +20,9 @@ public:
     void SetOnClose(std::function<void()> callback);
     void SetPosition(float x, float y);
 
+    /// v1.6: file drag-drop support
+    void SetDropCallback(std::function<void(std::vector<std::string>)> cb) { onDrop_ = std::move(cb); }
+
 private:
     std::string title_;
     std::vector<std::shared_ptr<Panel>> panels_;
@@ -27,6 +30,7 @@ private:
     float width_ = 0, height_ = 0;
     float pos_x_ = -1, pos_y_ = -1;
     std::function<void()> on_close_;
+    std::function<void(std::vector<std::string>)> onDrop_;
 };
 
 } // namespace unigui